image.bbclass: Update wrlt_next_class to support relative layers
Issue: LIN9-3600

Layers that are specified as a relative path from the build directory may
not patch properly and will cause a ValueError exception.  Fix this issue
by making sure that the bbpath being processed is always expanded.

If a layer is included more then once, we need to also skip to the last entry
or we will fail to load the next class in order.  (The result may not work
in all cases, however it won't work at all without the change.)

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>