layer.conf: set ESDK_MANIFEST_EXCLUDES
Issue: LIN10-2424

Set ESDK_MANIFEST_EXCLUDES to include wrtemplate* and wrimage* files.

These files will be refreshed at the first run of bitbake, and they
contain time information.

In the situation of eSDK update, as these files' checksums are changed,
the eSDK would always be updated, even if the remote eSDK is totally the
same with the local eSDK.

So we set ESDK_MANIFEST_EXCLUDES to avoid meaningless update.

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>