wrltemplate.bbclass: search bsp templates in parent layer
Issue: LIN10-2539

Search bsp templates in parent layer if it is a sublayer. For example:
$ ls intel-x86
intel-kabylake
templates

$ ls -l intel-x86/templates/default/bsp-pkgs.*
intel-x86/templates/default/bsp-pkgs.conf
intel-x86/templates/default/bsp-pkgs.inc

And there are several whitelist definations in bsp-pkgs.conf, e.g.:
PNWHITELIST_openembedded-layer += 'iperf3'

Since intel-kabylake is a sublayer of intel-x86, so it uses some common
configuration with the paraent layer, which brings in iperf3 here, and then
errors:

$ bitbake wrlinux-image-glibc-std-sato
lib32-iperf3 was skipped: Not supported in this configuration by Wind River. To override, add to your local.conf: PNWHITELIST_openembedded-layer += 'iperf3'
[snip]

The dependencies is: packagegroup-wr-bsps RDEPENDS on iperf3 when network is in
MACHINE_FEATURES, and conf/machine/intel-x86.inc adds network to
MACHINE_FEATURES, then iperf3 is pulled but it is not in whitelist since
intel-x86/templates/default/bsp-pkgs.conf is not handled, so we will get the
error.

Search bsp templates in parent layer can fix the problem.

Signed-off-by: Robert Yang <liezhi.yang@windriver.com>